[Issue-690] API for sample update refactor (#911)
* sdk refactoring

* test fixes;

* sdk refactoring

* small sdk fix; user generation cli cmd fix;

* code style fixes

* Update backend/client/deepchecks_client/core/client.py


* batch_update signature update

* fixes

* bug fix

* add handle of pytorch tensors to DeepchecksEncoder

+++ backend/tests/conftest.py
@@ -206,9 +206,12 @@ def settings(async_engine, smtp_server):
 
 @pytest.fixture(scope="session")
 def redis():
-    with subprocess.Popen(["redis-server", "--port", "6380"]) as p:
-        yield Redis(port=6380)
-        p.kill()
+    if (uri := os.environ.get("TESTS_REDIS_URI")) is not None:
+        yield Redis.from_url(uri)
+    else:
+        with subprocess.Popen(["redis-server", "--port", "6380"]) as p:
+            yield Redis(port=6380)
+            p.kill()
 
 
 @pytest.fixture(scope="function", autouse=True)
